Improved error handling

+++ rp_handler.py
@@ -295,11 +295,14 @@ def handler(event):
                         filename = image_filename['filename']
                         image_path = f'{VOLUME_MOUNT_PATH}/ComfyUI/output/{filename}'
 
-                        with open(image_path, 'rb') as image_file:
-                            images.append(base64.b64encode(image_file.read()).decode('utf-8'))
+                        if os.path.exists(image_path):
+                            with open(image_path, 'rb') as image_file:
+                                images.append(base64.b64encode(image_file.read()).decode('utf-8'))
 
-                        logging.info(f'Deleting output file: {image_path}', job_id)
-                        os.remove(image_path)
+                            logging.info(f'Deleting output file: {image_path}', job_id)
+                            os.remove(image_path)
+                        else:
+                            logging.error(f'Output file {image_path} not found')
 
                     return {
                         'images': images
